requirements.txt
django
djangorestframework
dynamicforms >= 0.74.22
swapper
django-rest-registration
drf-spectacular>=0.17.2, <0.26.0
django-taggit
svgwrite
django-hijack < 3
pytz
tox
social-auth-app-django
django-permissions-policy
django-csp
requests
boto3
natural
celery==5.2.7
django-redis-cache
pandas
dill==0.3.7
